Implemented Game::GetPlayer and object types.
The Game::GetPlayer function was properly implemented, and now we have a
(functional) 'type' enum inside every object to identify it, which (I
hope) will be useful when doing collision checking and strategy planning
for the enemy AI.

To be able to implement the object type system, I couldn't just use the
Object class constructor, so now all objects extend it, running their
own code after it. This allowed me not only to implement the object
types but also to make their constructors way cleaner. I'm very happy
with this commit :)