Working level event system
Had to do quite the overhaul, but now I have a functioning level event
system. Yay!

It reads and interprets events correctly. I still have unimplemented
events but hey, it now works! I decided to change LevelEvent from a
struct to an object so it'd be easier to add them to the events vector
and also to prevent memory leaks.

I added a test level to demo the level event system. All it does right
now is speed up to a crazy speed on line 160. I'm really happy with this
:)