Implemented (incomplete) slider enemy and enemy beam.
Also, I fixed LEVENT_SETBLINK and made the Beam constructor actually
save the beam type in the beamtype variable. I added an event that adds
a Slider in the testlevel. It's pretty cool so far!

Now I need to make sure that slider only fires when the player (rather,
either player) is within the randomly defined shot radius. After doing
that, I can try to add collisions to the game. (yaaay...)